rename cli sub-folder to cli-poc
