+++ application-operator/controllers/metricstrait/metricstrait_controller_test.go
@@ -2552,7 +2552,7 @@ func TestUseHTTPSForScrapeTargetTrueCondition(t *testing.T) {
 }
 
 // newMetricsTraitReconciler creates a new reconciler for testing
-// cli - The Kerberos client to inject into the reconciler
+// cli-poc - The Kerberos client to inject into the reconciler
 func newMetricsTraitReconciler(cli client.Client) Reconciler {
 	scheme := runtime.NewScheme()
 	vzapi.AddToScheme(scheme)
