+++ tools/cli-poc/scripts/cli_test.sh
