+++ tools/cli-poc/vz/cmd/cluster/get_registration_manifest.go
@@ -6,8 +6,9 @@ package cluster
 import (
 	"context"
 	"fmt"
+
 	"github.com/spf13/cobra"
-	"github.com/verrazzano/verrazzano/tools/cli/vz/pkg/helpers"
+	"github.com/verrazzano/verrazzano/tools/cli-poc/vz/pkg/helpers"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/cli-runtime/pkg/genericclioptions"
 )
