+++ tools/cli-poc/vz/cmd/root.go
@@ -11,12 +11,12 @@ import (
 	"github.com/verrazzano/verrazzano/pkg/k8sutil"
 	clustersclientset "github.com/verrazzano/verrazzano/platform-operator/clients/clusters/clientset/versioned"
 	verrazzanoclientset "github.com/verrazzano/verrazzano/platform-operator/clients/verrazzano/clientset/versioned"
-	"github.com/verrazzano/verrazzano/tools/cli/vz/cmd/app"
-	"github.com/verrazzano/verrazzano/tools/cli/vz/cmd/cluster"
-	"github.com/verrazzano/verrazzano/tools/cli/vz/cmd/login"
-	"github.com/verrazzano/verrazzano/tools/cli/vz/cmd/logout"
-	"github.com/verrazzano/verrazzano/tools/cli/vz/cmd/project"
-	"github.com/verrazzano/verrazzano/tools/cli/vz/pkg/helpers"
+	"github.com/verrazzano/verrazzano/tools/cli-poc/vz/cmd/app"
+	"github.com/verrazzano/verrazzano/tools/cli-poc/vz/cmd/cluster"
+	"github.com/verrazzano/verrazzano/tools/cli-poc/vz/cmd/login"
+	"github.com/verrazzano/verrazzano/tools/cli-poc/vz/cmd/logout"
+	"github.com/verrazzano/verrazzano/tools/cli-poc/vz/cmd/project"
+	"github.com/verrazzano/verrazzano/tools/cli-poc/vz/pkg/helpers"
 	"k8s.io/cli-runtime/pkg/genericclioptions"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
