+++ tools/cli-poc/vz/pkg/helpers/auth_status.go
