+++ tools/cli-poc/vz/pkg/helpers/browser.go
