+++ tools/cli-poc/vz/pkg/helpers/constants.go
