+++ tools/cli-poc/vz/pkg/helpers/print_flags.go
