+++ tools/cli-poc/vz/pkg/helpers/printer.go
