+++ tools/cli-poc/vz/pkg/helpers/printer_test.go
